pyproject.toml
[tool.poetry]
name = "logos-engine"
version = "0.1.0"
description = "Logos Framework Python engine (reference)"
authors = ["SolveForce <dev@solveforce.example>"]

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pyyaml = "^6.0"
openai = "^1.48.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
